// src/enums/routerEnum.ts
/*
 * @Description: Copyright (c) ydfk. All rights reserved
 * @Author: ydfk
 * @Date: 2021-08-26 12:43:31
 * @LastEditors: ydfk
 * @LastEditTime: 2021-08-28 21:01:15
 */
export enum RouterEnum {
  Error = "error",
  Home = "home",
  Login = "login",
  Account = "account",
  Statistic = "statistic",
  About = "about",
}